Switch from minidom to quick-xml
The minidom library was causing issues with the parsing, it wasn't
reading the junitxml input correctly and therefore none of the parsing
code was ever run. To address this issue this commit switches to using
the quick-xml library instead. The quick-xml library is what's used to
build minidom and was able to read the files without issue.

There is still a bug in the processing of the duration time. It's off by
several orders of magnitude, which likely means the power multipler
isn't the correct thing to do. This will be corrected in a later patch.