Fix test duration processing
The xml test
duration is a floating point value (which is parsed as a f64) while the
Duration constructor methods take in a i64. To avoid losing the
precision in the floating point before casting the f64 as an i64 we
multiplied it by 10^9 to use nanoseconds as an integer. However, the
code handling this neglected to use the Duration::nanoseconds()
constructor method and was instead passing the nanoseconds value to
Duration::seconds() constructor, leading to an very inflated value
in the output subunit. This commit fixes this issue by using the proper
nanoseconds() method to construct the Duration object.