Cleanup CLI UI
This commit makes cleans up the UI for using junitxml2subunit. It
switches to using the clap library to handle the argument processing and
building the CLI interface. By doing this the command gains a --help and
--version argument. Additionally some of the error handling is improved
by separating out different classes of read errors for the input XML and
given more descriptive error messages.