Fix bugs in test name handling
This commit fixes a few bugs in the test name handling. junitxml2subunit
was not properly handling several cases where the input xml had
different ways of describing the test name. This makes sure it is more
able to figure out how to uniquely identify the test_id for the subunit
output.