Handle text content for non success status tests
This commit adds support for handling Text content on tests that aren't
a success status. When parsing the xml input if there is text content
for a failed tag it should be treated as the error message (or reason
for a skipped test). However, junitxml2subunit wasn't handling this
case correctly and just ignoring that content.

To do this the failure and skip case handling had to be reworked
slightly because it's not possible to build a completed second packet
anymore from a single loop iteration. If there is a message found in the
attributes then it works as before. But, if no message is found then we
we have to wait until the next testcase event is encountered to truly
say we either did or didn't get a messsage. If we encountered a Text
event in between a status tag and the testcase tag that will be treated
as the message for the second packet from the status tag. This is
essentially how we were treating Text content successful tests before.