Properly handle send packet on xml EOF
When we were encountering an EOF event during the XML we were
unconditionally writing a second packet. However, that wasn't correct
because the normal case we don't need to write a packet. We should only
write the second packet there if the file ends and we're in between a
first and second packet waiting for an attachment. This commit fixes the
logic when handling an EOF to make sure we only write the second packet
when we need to.