Fix class and test name id creation.
When the test_id is being created for the subunit stream if there is
both a classname and a name (or id) attribute present the test_id for
the subunit stream is supposed to be a combination of the two. The
intent was to separate the 2 sections of the id by a '.', however
adding this period was missing. So the two components were just
concatenated together, resulting in a difficult to parse test_id. This
commit fixes this oversight and adds the missing period.