Add python testing for CLI interface
This commit adds python based tests to verify that the compiled output
command works as intended. While possible to build this testing natively
in rust, it was much faster and simpler to do so in python.
Additionally, it provides the added benefit of testing the
interoperability of the subunit output from the command. The subunit
protocol is language agnostic so verifying the subunit-rust library is
generated valid output by using a different library provides useful
validation.

The example JUnit XML files were either generated (in the case of the
pytest file) or downloaded by searching. The sources for these were
found at:

https://www.ibm.com/support/knowledgecenter/en/SSQ2R2_9.1.1/com.ibm.rsar.analysis.codereview.cobol.doc/topics/cac_useresults_junit.html

for the ant.xml example. The hudson example was based on:

https://github.com/notnoop/hudson-tools/blob/master/toJunitXML/sample-junit.xml

but with fixes for typos and XML syntax.