Fix deprecation warnings and remove deny warnings
In newer versions of the rustc compiler it's started emitting a
deprecation warning "trait objects without an explicit `dyn` are
deprecated". This was made fatal because deny warnings was set in
main.rs. This commit fixes the warnings by inserting the dyn keyword
as suggested by the compiler. Then to prevent this issue from occuring
in the future deny warnings is removed from the file.