Pivot CI to github actions
With the tight travis quotas for open source projects now it is not a
viable provider for running CI on this project anymore. While this
project is relatively low activity (as it's mostly stable at this point)
the quota is shared on my github account and other projects eat all the
availble quota. This commit switches CI to use github actions which
doesn't have this limitation.